+++ config/bat.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'admin_emails' => array_filter(explode(',', env('BAT_ADMIN_EMAILS', ''))),
-];