+++ resources/views/bat/show.blade.php
@@ -1,3 +1,0 @@
-<x-layouts.app title="Bon a Tirer - {{ $token }}">
-    <livewire:bat-approval :token="$token" />
-</x-layouts.app>